requirements.txt
flask
flask_cors
mysql_connector
werkzeug